requirements.txt
django==4.2.2
python-decouple==3.8
psycopg2-binary==2.9.6
django-cors-headers==4.0.0
djangorestframework==3.14.0
drf_spectacular==0.26.2
drf-spectacular-sidecar==2023.6.1
openai==0.27.8